requirements.txt
tornado==2.3
pyzmq==2.2.0
-e git+git://github.com/mrjoes/sockjs-tornado.git#egg=sockjs
